finish

+++ services/product.js
@@ -1,7 +1,7 @@
 const Product = require("../model/product");
 const User = require("../model/user");
 
-exports.handleGetAllProduct = async(limit, page) => {
+exports.handleGetAllProduct = async (limit, page) => {
   return new Promise(async (resolve, reject) => {
     try {
       const products = await Product.find();
@@ -32,18 +32,23 @@ exports.handleGetAllProduct = async(limit, page) => {
 exports.handleAddProduct = (value) => {
   return new Promise(async (resolve, reject) => {
     try {
-      const product = new Product({
-        name: value.name,
-        price: value.price,
-        short_desc: value.shortDesc,
-        long_desc: value.longDesc,
-        category: value.category,
-        count: value.count,
-        images: value.files,
-      });
-      const newProduct = await product.save();
-      if (newProduct) {
-        resolve({ statusCode: 200, message: "ok" });
+      const user = await User.findById(value.userId);
+      if (user.role === 2) {
+        const product = new Product({
+          name: value.name,
+          price: value.price,
+          short_desc: value.shortDesc,
+          long_desc: value.longDesc,
+          category: value.category,
+          count: value.count,
+          images: value.files,
+        });
+        const newProduct = await product.save();
+        if (newProduct) {
+          resolve({ statusCode: 200, message: "ok" });
+        }
+      } else {
+        resolve({ statusCode: 403, message: "Unauthorized" });
       }
     } catch (err) {
       reject(err);
@@ -54,19 +59,24 @@ exports.handleAddProduct = (value) => {
 exports.handleUpdateProduct = (value) => {
   return new Promise(async (resolve, reject) => {
     try {
-      const product = await Product.findById(value.productId);
-      if (product) {
-        product.name = value.name;
-        product.price = value.price;
-        product.short_desc = value.shortDesc;
-        product.long_desc = value.longDesc;
-        product.category = value.category;
-        product.count = value.count;
-        product.images = value.files;
-        const newProduct = await product.save();
-        if (newProduct) {
-          resolve({ statusCode: 200, message: "ok" });
+      const user = await User.findById(value.userId);
+      if (user.role === 2) {
+        const product = await Product.findById(value.productId);
+        if (product) {
+          product.name = value.name;
+          product.price = value.price;
+          product.short_desc = value.shortDesc;
+          product.long_desc = value.longDesc;
+          product.category = value.category;
+          product.count = value.count;
+          product.images = value.files;
+          const newProduct = await product.save();
+          if (newProduct) {
+            resolve({ statusCode: 200, message: "ok" });
+          }
         }
+      } else {
+        resolve({ statusCode: 403, message: "Unauthorized" });
       }
     } catch (err) {
       reject(err);
@@ -74,28 +84,33 @@ exports.handleUpdateProduct = (value) => {
   });
 };
 
-exports.handleDeleteProduct = (productId) => {
-  return new Promise(async(resolve, reject) => {
-    try{
-      const product = await Product.findByIdAndDelete(productId);
-      if(product) {
-        resolve({ message: 'ok', statusCode: 200 });
+exports.handleDeleteProduct = (productId, userId) => {
+  return new Promise(async (resolve, reject) => {
+    try {
+      const user = await User.findById(userId);
+      if (user.role === 2) {
+        const product = await Product.findByIdAndDelete(productId);
+        if (product) {
+          resolve({ message: "ok", statusCode: 200 });
+        }
+      } else {
+        resolve({ statusCode: 403, message: "Unauthorized" });
       }
-    }catch(err) {
+    } catch (err) {
       reject(err);
     }
-  })
+  });
 };
 
 exports.handleGetEditProduct = (productId) => {
-  return new Promise(async(resolve, reject) => {
-    try{
+  return new Promise(async (resolve, reject) => {
+    try {
       const product = await Product.findById(productId);
-      if(product) {
-        resolve({ message: 'ok', statusCode: 200, product: product });
+      if (product) {
+        resolve({ message: "ok", statusCode: 200, product: product });
       }
-    }catch(err) {
+    } catch (err) {
       reject(err);
     }
-  })
-}
+  });
+};
